requirements.txt
cachetools==2.0.1
certifi==2017.11.5
cffi==1.11.2
chardet==3.0.4
click==6.7
futures==3.1.1
google-assistant-grpc==0.1.0
google-assistant-sdk==0.4.0
google-auth==1.2.1
google-auth-oauthlib==0.2.0
googleapis-common-protos==1.5.3
grpcio==1.8.2
idna==2.6
monotonic==1.4
oauthlib==2.0.6
protobuf==3.5.1
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycparser==2.18
requests==2.18.4
requests-oauthlib==0.8.0
rsa==3.4.2
six==1.11.0
sounddevice==0.3.10
tenacity==4.8.0
urllib3==1.22
